requirements.txt
anthropic>=0.39.0
pydantic>=2.5.0
pandas>=2.1.0
openpyxl>=3.1.0
python-docx>=1.1.0
pyyaml>=6.0
python-dotenv>=1.0.0
rich>=13.7.0
click>=8.1.0
streamlit>=1.30.0
PyMuPDF>=1.24.0
pdfplumber>=0.11.0
